Rework SVG renderer

+++ src/HttpgdApi.h
@@ -17,7 +17,7 @@ namespace httpgd
         virtual bool api_clear() = 0;
 
         virtual void api_svg(std::ostream &os, int index, double width, double height) = 0;
-        virtual boost::optional<int> api_index(long id) = 0;
+        virtual boost::optional<int> api_index(int32_t id) = 0;
 
         virtual HttpgdState api_state() = 0;
